feat: redeploys amoy with latest changes and finishs deployment on sepolia
