+++ scripts/ccip/syncVotes.ts
@@ -13,14 +13,14 @@ import { getRpc } from "./rpcs";
 async function main() {
   const alchemyKey = process.env.ALCHEMY_RPC_KEY ?? "";
   const sourceChainName = "sepolia";
-  const targetChainName = process.argv[2] ?? "";
-  const voters = process.argv[3].split(",");
+  const senderAddress = process.argv[2] ?? "";
+  const targetChainName = process.argv[3] ?? "";
+  const targetAddress = process.argv[4] ?? "";
+  const voters = process.argv[5].split(",");
   const sourceProvider = new ethers.JsonRpcProvider(
     getRpc(sourceChainName, alchemyKey)
   );
-  const targetProvider = new ethers.JsonRpcProvider(
-    getRpc(targetChainName, alchemyKey)
-  );
+
   console.log(`Sending votes to ${targetChainName}`);
   const sourceWallet = ethers.Wallet.fromPhrase(
     process.env.DEPLOYER_ACCOUNT_SEED ?? "",
@@ -31,90 +31,39 @@ async function main() {
     (param) => param.chainName.toLowerCase() === targetChainName.toLowerCase()
   );
 
-  const senderFile = await fs.readFile(
-    `scripts/deployment/deployments/${sourceChainName}/GovernanceSender.json`
-  );
-  const senderParsed = JSON.parse(senderFile.toString());
-  console.log(`Using GovernanceSender ${senderParsed.address}`);
+  console.log(`Using GovernanceSender ${senderAddress}`);
 
-  const receiverFile = await fs.readFile(
-    `scripts/deployment/deployments/${targetChainName}/BridgedGovernance.json`
-  );
-  const receiverParsed = JSON.parse(receiverFile.toString());
   console.log(
-    `Using BridgedGovernance (${receiverParsed.address}) on ${targetChainName} `
+    `Using BridgedGovernance (${targetAddress}) on ${targetChainName} `
   );
 
   const governanceSender = GovernanceSender__factory.connect(
-    senderParsed.address,
+    senderAddress,
     sourceWallet
   );
-  const bridgedGovernance = BridgedGovernance__factory.connect(
-    receiverParsed.address,
-    targetProvider
-  );
-  const ccipMessage = await governanceSender.getCCIPMessage(
-    receiverParsed.address,
-    ethers.ZeroAddress,
+
+  const fee = await governanceSender["getSyncFee(uint64,address,address[],bool)"](
+    targetParams?.chainSelector ?? '',
+    targetAddress,
     voters,
-    "0x"
-  );
-  const receiveFunc = bridgedGovernance.getFunction("ccipReceive");
-  const receiveTx = await receiveFunc.populateTransaction({
-    messageId:
-      "0x2ccc749c286b1effdf3df51fae94792e1f10b29fbd8a7dc97095448a24b153ab",
-    sourceChainSelector: await bridgedGovernance.MAINNET_CHAIN_SELECTOR(),
-    sender: ethers.AbiCoder.defaultAbiCoder().encode(
-      ["address"],
-      [await bridgedGovernance.MAINNET_GOVERNANCE_ADDRESS()]
-    ),
-    data: ccipMessage.data,
-    destTokenAmounts: ccipMessage.tokenAmounts,
-  });
-  receiveTx.from = await bridgedGovernance.getRouter();
-  const gasEstimation = await targetProvider.estimateGas(receiveTx);
-  console.log(`Estimated receive uses ${gasEstimation} gas`);
+    true
+  )
 
-  const functionSelector = ethers.id("CCIP EVMExtraArgsV2").slice(0, 10);
-  const extraArgs =
-    functionSelector +
-    ethers.AbiCoder.defaultAbiCoder()
-      .encode(["(uint256, bool)"], [[gasEstimation, true]])
-      .slice(2);
-
-  console.log(
-    `Getting ccip fee with 
-      ${receiverParsed.address} 
-      ${targetParams?.chainSelector ?? ""} 
-      ${ethers.ZeroAddress} 
-      ${voters} 
-      ${extraArgs} 
-    `
-  );
-
-  const ccipFee = await governanceSender.getCCIPFee(
-    receiverParsed.address,
-    targetParams?.chainSelector ?? "",
-    ethers.ZeroAddress,
+  console.log(`Fee: ${fee.toString()} in native`);
+  const tx = await governanceSender["syncVotes(uint64,address,address[])"](
+    targetParams?.chainSelector ?? '',
+    targetAddress,
     voters,
-    extraArgs
-  );
-  console.log(`CCIP Fee: ${ccipFee}`);
-
-  const syncTx = await governanceSender.syncVotesPayNative(
-    ethers.AbiCoder.defaultAbiCoder().encode(
-      ["address"],
-      [receiverParsed.address]
-    ),
-    targetParams?.chainSelector ?? "",
-    voters,
-    extraArgs,
     {
-      value: ccipFee,
+      value: fee,
     }
-  );
-  console.log(`Sent tx ${syncTx.hash}. Waiting for validators...`);
-  await syncTx.wait();
+  )
+  console.log(`Sent transaction ${tx.hash}`);
+  const receipt = await tx.wait();
+  if(!receipt) {
+    throw new Error("Transaction failed");
+  }
+  console.log(`Transaction confirmed on ${receipt.blockNumber}`);
 }
 
 main();
