+++ scripts/deployment/parameters/paramsCCIP.json
@@ -20,8 +20,8 @@
     "mainnetGovernanceSender": "0x08AdafD51205cEB550BB1b34769E89c8A609e6cA",
     "rmnProxy": "0x7c1e545A40750Ee8761282382D51E017BAC68CBB",
     "registryModuleOwner": "0x84ad5890A63957C960e0F19b0448A038a574936B",
-    "zchf": "0x8BBcbe81FFCa30DE315d4CF6fdF839F7A01Aa4E3",
-    "tokenPool": "0x763fE4f60528dbc560C206030faFB4Ab883D4d3E",
+    "zchf": "0xB6D2E9998C194e627Db2FcBa176313c35BFE52E8",
+    "tokenPool": "0x59dB29840a092F43Df2D74514e5419Af2D4F4C8d",
     "link": "0x0Fd9e8d3aF1aaee056EB9e802c3A762a667b1904",
     "bridgeAccounting": "0x3bF23a52590C2DF855bDbA748489C5fbBA54da12",
     "mainnetLeadrateSender": "0xB0351e929F6Bd42873Ec795919E0f03DbD797caa",
